:truck: Rename ActiveRecord::Dbt::Factory::Model::StagingFactory to ActiveRecord::Dbt::Factory::Model::Staging::YmlFactory
